remove no longer necessary nesting of modules on `compat` utils
