+++ mev-build-rs/src/lib.rs
@@ -1,10 +1,10 @@
 mod auctioneer;
 mod bidder;
+mod compat;
 mod error;
 mod node;
 mod payload;
 mod service;
-mod utils;
 
 pub use crate::error::Error;
 pub use service::{launch, Config};
